# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ocha URL Shorten Service",
        "title": "Ocha URL Shorten Service",
        "contact": {
            "name": "rahulvramesh",
            "email": "rahulvramesh@hotmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/shorten": {
            "post": {
                "description": "Create short url handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create short url handler",
                "operationId": "create-short-url",
                "parameters": [
                    {
                        "description": "New Short URL",
                        "name": "short-url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "models.CreateRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "shortcode": {
                    "type": "string",
                    "example": "google"
                },
                "url": {
                    "type": "string",
                    "example": "https://www.google.com"
                }
            }
        },
        "models.CreateResponse": {
            "type": "object",
            "properties": {
                "shortcode": {
                    "type": "string",
                    "example": "google"
                }
            }
        }
    }
}